If there is no hostname for route, create a default
If a host for a route is set to "", then the router will generate
a default host name automatically.
If the host is set to "something" but then changed to "", the previous
host value ("something") will remain, the route controller will
not set host back to the generated default.

Resetting back to the generated value can be emulated by working out
what the generated value should be, and explicitly setting host to
this value